The modal.js script was relocated to the "./js" folder
